Retain dtype of fields during remesh_align
In my org's usage of POPY, we store `num_samples`,
`total_sample_weight`, and others as `np.float32` in order to save
storage and memory when producing large Level 3 products.

This change retains the `dtype` of the existing fields. Without it,
fields of type `np.float32` would come out of `remesh_align` as
`np.float64`.